src/test/java/seedu/momentum/storage/JsonSerializableProjectBookTest.java
package seedu.momentum.storage;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static seedu.momentum.testutil.Assert.assertThrows;

import java.nio.file.Path;
import java.nio.file.Paths;

import org.junit.jupiter.api.Test;

import seedu.momentum.commons.exceptions.IllegalValueException;
import seedu.momentum.commons.util.JsonUtil;
import seedu.momentum.model.ProjectBook;
import seedu.momentum.testutil.TypicalPersons;

public class JsonSerializableProjectBookTest {

    private static final Path TEST_DATA_FOLDER = Paths.get("src", "test", "data", "JsonSerializableProjectBookTest");
    private static final Path TYPICAL_PERSONS_FILE = TEST_DATA_FOLDER.resolve("typicalPersonsProjectBook.json");
    private static final Path INVALID_PERSON_FILE = TEST_DATA_FOLDER.resolve("invalidPersonProjectBook.json");
    private static final Path DUPLICATE_PERSON_FILE = TEST_DATA_FOLDER.resolve("duplicatePersonProjectBook.json");

    @Test
    public void toModelType_typicalPersonsFile_success() throws Exception {
        JsonSerializableProjectBook dataFromFile = JsonUtil.readJsonFile(TYPICAL_PERSONS_FILE,
                JsonSerializableProjectBook.class).get();
        ProjectBook projectBookFromFile = dataFromFile.toModelType();
        ProjectBook typicalPersonsProjectBook = TypicalPersons.getTypicalProjectBook();
        assertEquals(projectBookFromFile, typicalPersonsProjectBook);
    }

    @Test
    public void toModelType_invalidPersonFile_throwsIllegalValueException() throws Exception {
        JsonSerializableProjectBook dataFromFile = JsonUtil.readJsonFile(INVALID_PERSON_FILE,
                JsonSerializableProjectBook.class).get();
        assertThrows(IllegalValueException.class, dataFromFile::toModelType);
    }

    @Test
    public void toModelType_duplicatePersons_throwsIllegalValueException() throws Exception {
        JsonSerializableProjectBook dataFromFile = JsonUtil.readJsonFile(DUPLICATE_PERSON_FILE,
                JsonSerializableProjectBook.class).get();
        assertThrows(IllegalValueException.class, JsonSerializableProjectBook.MESSAGE_DUPLICATE_PERSON,
                dataFromFile::toModelType);
    }

}
